Add `zlib-helper` build to packager.
- Note: Right now only supports zlib build from mono tree build.
- Add `zlibtest.cs` scenario to compress and uncompress using DeflateStream and GZip.
- Add test to make file to build and run aot version of ziplibtest.

+++ sdks/wasm/zlib-helper.c
@@ -329,7 +329,7 @@ WriteZStream (ZStream *stream, guchar *buffer, gint length)
 MonoMethod* 
 GetReadOrWriteCallback (gint compress, void *gchandle)
 {
-	MonoObject* callback = mono_gchandle_get_target (gchandle);
+	MonoObject* callback = mono_gchandle_get_target ((guint32)gchandle);
 	if (!callback) {
 		fprintf(stderr, "Error: zlib-helper - Callback target not found.\n");
 		return NULL;
